refactor postdata changes into separate class
